refactor(backend): split responses transformer into submodules
